requirements.txt
PySide6>=6.6
PySide6-Fluent-Widgets>=1.9.0
Pillow>=10.0.0
fonttools>=4.40.0
